requirements.txt
streamlit==1.32.0
flickrapi==2.4.0
Pillow==10.2.0
numpy==1.24.3
opencv-python-headless==4.8.0.76
requests==2.31.0
ultralytics
pyvirtualdisplay==3.0